chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.111.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchainhub>=0.1.16",
    "langchain_community>=0.0.30",
    "numpy>=1.26.4",
    "openai>=1.30.5",
    "opentelemetry-api>=1.24.0",
    "pydantic>=2.7.1",
    "uvicorn>=0.29.0",
    "python-dotenv>=1.0.1",
    "bs4>=0.0.2"
]